Move PositionComparator to SortedPositionLinks
