+++ presto-main/src/main/java/com/facebook/presto/operator/JoinHashSupplier.java
@@ -16,7 +16,6 @@ package com.facebook.presto.operator;
 import com.facebook.presto.Session;
 import com.facebook.presto.spi.block.Block;
 import com.facebook.presto.sql.gen.JoinFilterFunctionCompiler.JoinFilterFunctionFactory;
-import it.unimi.dsi.fastutil.ints.IntComparator;
 import it.unimi.dsi.fastutil.longs.LongArrayList;
 
 import java.util.List;
@@ -24,8 +23,6 @@ import java.util.Optional;
 import java.util.function.Function;
 
 import static com.facebook.presto.SystemSessionProperties.isFastInequalityJoin;
-import static com.facebook.presto.operator.SyntheticAddress.decodePosition;
-import static com.facebook.presto.operator.SyntheticAddress.decodeSliceIndex;
 import static java.util.Objects.requireNonNull;
 
 public class JoinHashSupplier
@@ -57,7 +54,6 @@ public class JoinHashSupplier
                 isFastInequalityJoin(session)) {
             positionLinksBuilder = SortedPositionLinks.builder(
                     addresses.size(),
-                    new PositionComparator(pagesHashStrategy, addresses),
                     pagesHashStrategy,
                     addresses);
         }
@@ -92,38 +88,5 @@ public class JoinHashSupplier
                 pagesHash,
                 filterFunction,
                 positionLinks.apply(filterFunction));
-    }
-
-    public static class PositionComparator
-            implements IntComparator
-    {
-        private final PagesHashStrategy pagesHashStrategy;
-        private final LongArrayList addresses;
-
-        public PositionComparator(PagesHashStrategy pagesHashStrategy, LongArrayList addresses)
-        {
-            this.pagesHashStrategy = pagesHashStrategy;
-            this.addresses = addresses;
-        }
-
-        @Override
-        public int compare(int leftPosition, int rightPosition)
-        {
-            long leftPageAddress = addresses.getLong(leftPosition);
-            int leftBlockIndex = decodeSliceIndex(leftPageAddress);
-            int leftBlockPosition = decodePosition(leftPageAddress);
-
-            long rightPageAddress = addresses.getLong(rightPosition);
-            int rightBlockIndex = decodeSliceIndex(rightPageAddress);
-            int rightBlockPosition = decodePosition(rightPageAddress);
-
-            return pagesHashStrategy.compareSortChannelPositions(leftBlockIndex, leftBlockPosition, rightBlockIndex, rightBlockPosition);
-        }
-
-        @Override
-        public int compare(Integer leftPosition, Integer rightPosition)
-        {
-            return compare(leftPosition.intValue(), rightPosition.intValue());
-        }
     }
 }
